+++ src/kitaru/_default_plugins/__init__.py
@@ -1,4 +1,0 @@
-#  Copyright (c) ZenML GmbH 2026. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-"""Source resources for Kitaru's released default plugins."""